requirements.txt
fastapi==0.115.5
fastapi-mail==1.4.0
passlib==1.7.4
psycopg2==2.9.10
# FOR MACOS-LINUX USE BINARY INSTEAD
# psycopg2-binary==2.9.10
pydantic==2.10.1
python-dotenv==1.0.1
python_jose==3.3.0
sqlalchemy==2.0.36

